change how users appear
